update: - move antlr def to this - break cursor action into smaller classes
