+++ buildLib.sh
@@ -1,4 +1,5 @@
-./gradlew :err:clean :err:build :err:publishToMavenLocal
-./gradlew :common:clean :common:build :common:publishToMavenLocal
-./gradlew :common:clean :common:build :common-compose:publishToMavenLocal
-./gradlew :err:clean :err:build :err:publishToMavenLocal
+#./gradlew :err:clean :err:build :err:publishToMavenLocal
+#./gradlew :common:clean :common:build :common:publishToMavenLocal
+#./gradlew :common:clean :common:build :common-compose:publishToMavenLocal
+#./gradlew :err:clean :err:build :err:publishToMavenLocal
+./gradlew :p6-antlr:clean :p6-antlr:build :p6-antlr:publishToMavenLocal
