+++ common/build.gradle.kts
@@ -2,9 +2,6 @@
 
 plugins {
     java
-//    kotlin("jvm") version "1.6.10"
-//    kotlin("kapt") version "1.6.10"
-
     kotlin("jvm") version "1.7.20"
     kotlin("kapt") version "1.7.20"
     id("maven-publish")
