+++ p6-app/src/main/kotlin/com/qxdzbc/p6/app/action/cursor/copy_cursor_range_to_clipboard/CopyCursorRangeToClipboardAction.kt
@@ -1,0 +1,10 @@
+package com.qxdzbc.p6.app.action.cursor.copy_cursor_range_to_clipboard
+
+import com.qxdzbc.p6.app.action.common_data_structure.WbWs
+
+interface CopyCursorRangeToClipboardAction{
+    /**
+     * Compute a range from a cursor's state (often the main range), and copy it to clipboard
+     */
+    fun copyCursorRangeToClipboard(wbws: WbWs)
+}
