+++ p6-app/src/main/kotlin/com/qxdzbc/p6/app/action/cursor/paste_range_to_cursor/PasteRangeToCursor.kt
@@ -1,0 +1,10 @@
+package com.qxdzbc.p6.app.action.cursor.paste_range_to_cursor
+
+import com.qxdzbc.p6.app.action.common_data_structure.WbWs
+
+interface PasteRangeToCursor{
+    /**
+     * Paste whatever in the clipboard into the current position of a cursor at [wbws]
+     */
+    fun pasteRange(wbws: WbWs)
+}
