+++ p6-app/src/main/kotlin/com/qxdzbc/p6/app/action/cursor/thumb/drag_thumb_action/DragThumbAction.kt
@@ -1,4 +1,4 @@
-package com.qxdzbc.p6.ui.document.worksheet.cursor.thumb.action
+package com.qxdzbc.p6.app.action.cursor.thumb.drag_thumb_action
 
 import androidx.compose.ui.geometry.Offset
 import com.qxdzbc.p6.app.action.common_data_structure.WbWsSt
