+++ p6-app/src/main/kotlin/com/qxdzbc/p6/app/action/cursor/thumb/drag_thumb_action/EndThumbDragActionImp.kt
@@ -1,4 +1,4 @@
-package com.qxdzbc.p6.ui.document.worksheet.cursor.thumb.action
+package com.qxdzbc.p6.app.action.cursor.thumb.drag_thumb_action
 
 import androidx.compose.runtime.getValue
 import com.qxdzbc.common.compose.St
