+++ p6-app/src/main/kotlin/com/qxdzbc/p6/ui/document/worksheet/cursor/actions/CursorAction.kt
@@ -2,50 +2,35 @@ package com.qxdzbc.p6.ui.document.worksheet.cursor.actions
 
 import androidx.compose.ui.graphics.Color
 import com.qxdzbc.p6.app.action.common_data_structure.WbWs
-import com.qxdzbc.p6.app.common.key_event.P6KeyEvent
 import com.qxdzbc.p6.app.document.range.address.RangeAddress
+import com.qxdzbc.p6.ui.app.cell_editor.actions.CellEditorAction
+import com.qxdzbc.p6.app.action.cursor.copy_cursor_range_to_clipboard.CopyCursorRangeToClipboardAction
+import com.qxdzbc.p6.app.action.cursor.handle_cursor_keyboard_event.HandleCursorKeyboardEventAction
+import com.qxdzbc.p6.app.action.cursor.paste_range_to_cursor.PasteRangeToCursor
+import com.qxdzbc.p6.app.action.cursor.undo_on_cursor.UndoOnCursorAction
 import com.qxdzbc.p6.ui.document.worksheet.cursor.state.CursorStateId
+import com.qxdzbc.p6.ui.document.worksheet.cursor.thumb.action.ThumbAction
 import com.qxdzbc.p6.ui.document.worksheet.select_whole_col_for_selected_cell.SelectWholeColumnForAllSelectedCellAction
 import com.qxdzbc.p6.ui.document.worksheet.select_whole_row_for_selected_cells.SelectWholeRowForAllSelectedCellAction
 
+interface CursorAction :
+    SelectWholeColumnForAllSelectedCellAction,
+    SelectWholeRowForAllSelectedCellAction,
+    HandleCursorKeyboardEventAction,
+    PasteRangeToCursor,
+    CopyCursorRangeToClipboardAction,
+    UndoOnCursorAction
+{
+    val cellEditorAction: CellEditorAction
+    val thumbAction: ThumbAction
 
-interface CursorAction : SelectWholeColumnForAllSelectedCellAction, SelectWholeRowForAllSelectedCellAction {
+    fun moveCursorTo(wbws: WbWs, cellLabel: String)
 
-    /**
-     * f2 key
-     */
-    fun f2(wbws: WbWs)
-    fun home(wbws: WbWs)
-    fun end(wbws: WbWs)
-    fun ctrlUp(wbws: WbWs)
-    fun ctrlDown(wbws: WbWs)
-    fun ctrlRight(wbws: WbWs)
-    fun ctrlShiftLeft(wbws: WbWs)
-    fun ctrlShiftRight(wbws: WbWs)
-    fun ctrlShiftUp(wbws: WbWs)
-    fun ctrlShiftDown(wbws: WbWs)
-    fun ctrlLeft(wbws: WbWs)
-    fun up(wbws: WbWs)
-    fun down(wbws: WbWs)
-    fun left(wbws: WbWs)
-    fun right(wbws: WbWs)
-    fun moveCursorTo(wbws: WbWs,cellLabel:String)
-    fun shiftUp(wbws: WbWs)
-    fun shiftDown(wbws: WbWs)
-    fun shiftLeft(wbws: WbWs)
-    fun shiftRight(wbws: WbWs)
-
-    fun onDeleteKey(wbws: WbWs)
-    fun undo(wbws: WbWs)
-    fun handleKeyboardEvent(keyEvent: P6KeyEvent, wbws: WbWs): Boolean
-    fun pasteRange(wbws: WbWs)
-    fun rangeToClipboard(wbws: WbWs)
-
-    fun getFormulaRangeAndColor(wbws: WbWs):Map<RangeAddress,Color>
+    fun getFormulaRangeAndColor(wbws: WbWs): Map<RangeAddress, Color>
 
     fun focusOnCursor(cursorId: CursorStateId)
     fun freeFocusOnCursor(cursorId: CursorStateId)
-    fun updateCursorFocus(cursorId: CursorStateId,focused: Boolean)
+    fun updateCursorFocus(cursorId: CursorStateId, focused: Boolean)
 //    fun updateCellEditorFocus(cursorId: CursorStateId,focused: Boolean)
 }
