+++ p6-app/src/main/kotlin/com/qxdzbc/p6/ui/document/worksheet/cursor/thumb/action/ThumbAction.kt
@@ -1,5 +1,7 @@
 package com.qxdzbc.p6.ui.document.worksheet.cursor.thumb.action
 
+import com.qxdzbc.p6.app.action.cursor.thumb.drag_thumb_action.DragThumbAction
+
 
 interface ThumbAction : DragThumbAction {
