+++ p6-app/src/main/kotlin/com/qxdzbc/p6/ui/document/worksheet/cursor/thumb/action/ThumbActionImp.kt
@@ -2,6 +2,7 @@ package com.qxdzbc.p6.ui.document.worksheet.cursor.thumb.action
 
 import com.qxdzbc.p6.di.P6Singleton
 import com.qxdzbc.p6.di.anvil.P6AnvilScope
+import com.qxdzbc.p6.app.action.cursor.thumb.drag_thumb_action.DragThumbAction
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 @P6Singleton
