+++ p6-app/src/test/kotlin/com/qxdzbc/p6/ui/document/worksheet/cursor/thumb/action/DragThumbActionImpTest.kt
@@ -3,13 +3,13 @@ package com.qxdzbc.p6.ui.document.worksheet.cursor.thumb.action
 import com.qxdzbc.common.compose.Ms
 import com.qxdzbc.p6.app.action.common_data_structure.WbWs
 import com.qxdzbc.p6.app.action.common_data_structure.WbWsSt
-import com.qxdzbc.p6.app.document.cell.address.CellAddress
 import com.qxdzbc.p6.ui.document.worksheet.cursor.state.CursorState
+import com.qxdzbc.p6.app.action.cursor.thumb.drag_thumb_action.DragThumbAction
 import test.BaseTest
 import kotlin.test.*
 
 internal class DragThumbActionImpTest:BaseTest(){
-    lateinit var act:DragThumbAction
+    lateinit var act: DragThumbAction
     lateinit var wbwsSt:WbWsSt
     lateinit var cursorStateMs:Ms<CursorState>
     @BeforeTest
