+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include("common-compose","p6-app")
 val p6Version by extra("1.0")
 include("err")
 include("common")
+include("p6-antlr")
